LayoutTests/crypto/webkitSubtle/unwrapKey-check-usage-expected.txt

CONSOLE MESSAGE: line 40: WebKitSubtleCrypto is deprecated. Please use SubtleCrypto instead.
CONSOLE MESSAGE: line 1: Key usages do not include 'unwrapKey'
Verify that usage is respected in unwrapKey operation.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Importing RSA keys...
Importing an AES key...
Wrapping it...
Unwrapping it...
PASS crypto.webkitSubtle.unwrapKey("raw", wrappedKey, privateKey, wrapAlgorithm, "aes-cbc", extractable, ["encrypt", "decrypt"]) threw exception NotSupportedError (DOM Exception 9): The operation is not supported..
PASS successfullyParsed is true

TEST COMPLETE
